apps/api/pyproject.toml
[project]
name = "launchclaw-api"
version = "0.1.0"
description = "FastAPI control-plane service for LaunchClaw"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "croniter>=3.0,<4.0",
  "fastapi>=0.110,<1.0",
  "pydantic-settings>=2.2,<3.0",
  "supabase>=2.28.0",
  "uvicorn[standard]>=0.30,<1.0",
]

[tool.uv]
package = false
